Complete blog

+++ app/Http/Controllers/ProfileController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\UpdateProfileRequest;
 use App\Profile;
 use Illuminate\Support\Facades\Storage;
 
@@ -13,29 +14,25 @@ class ProfileController extends Controller
         return view('user.profile', compact('profile'));
     }
 
-    public function update(Profile $profile)
+    public function update(UpdateProfileRequest $request ,Profile $profile)
     {
-        $campos = request()->validate([
-            //'avatar' => 'image:jpeg,png,gif,svg|max:5120',
-            'about' => 'required|min:20',
-            'facebook' => 'required|url',
-            'youtube' => 'required|url'
-        ]);
-
-
-        /*if(request()->hasFile('avatar'))
-        {
-            try{
-                Storage::delete(auth()->user()->profile->avatar);
-            }catch (\Exception $e){
-                //
-            }
-            $campos['avatar'] = request()->file('avatar')->store('avatars', 'public');
-        }*/
+        $campos = $request->validated();
 
         $profile->fill($campos);
 
         auth()->user()->profile()->save($profile);
+
+        if(request()->hasFile('avatar'))
+        {
+            try{
+                Storage::delete(auth()->user()->profile->avatar);
+
+                tap(request()->file('avatar')->store('avatars', 'public'), function ($ruta) {
+                    auth()->user()->fill([ 'avatar' => $ruta])->save();
+                });
+
+            }catch (\Exception $e){}
+        }
 
         return back()->with(['success' => 'Account profile has been updated successfully.']);
     }
